Allow for a path-based post URL publishing scheme.
This change adds a prompt question to the `delposto create` sequence; it
asks whether or not the user wants the default date-based post URL
scheme, YYYY/MM/post-slug. The answer is written to the site's meta.json
and used in the `publish` process (I left `preview` untouched).

If the date-based URL scheme is turned off, posts' URLs will be created
based on the path to the draft being published, so:

`delposto draft something` results in the default behavior, but with
the post being published as /something/index.html.

`delposto draft some/nested/thing` results in the post being published
as /some/nested/thing/index.html.

The trailing slash behavior for the `draft` command is preserved, so
`delposto draft some/nested/thing/` creates a file
`drafts/some/nested/thing/index.md`, and you can put other files next
to it so they all get published together in
`/some/nested/thing/slugged-post-title/`.

The cleanup done after publishing a post is still limited to removal
of the post file or the post directory itself (for a trailing-slashed
draft). Nested directories aren't removed, ostensibly because the
assumption is that if you've nested a draft, you'll likely have other
posts/drafts in the same directory structure at some point. In
reality, this is simply a case of developer laziness on my part, but
I think I can get away with making the other claim sound legit :-)